Create pepper_map_v5_v2
Compared to the original pepper_map_v5 environment program, the original program would encounter errors due to the removal of the rendering syntax in the newer versions of gym. This version of the code replaces the original rendering functionality with matplotlib, allowing it to work with gym versions 0.22.0 and above.